Move logged_artifact and example_files fixtures to conftest.py
